fixed items from crit from Mark

+++ kanban.server/server/controllers/BoardsController.js
@@ -20,7 +20,8 @@ export class BoardsController extends BaseController {
 
   async getAllBoards(req, res, next) {
     try {
-      const data = await boardsService.getAllBoards(req.params.id)
+      //
+      const data = await boardsService.getAllBoards({ creatorId: req.userInfo.id })
       return res.send(data)
     } catch (error) {
       next(error)
@@ -39,7 +40,7 @@ export class BoardsController extends BaseController {
 
   async deleteBoard(req, res, next) {
     try {
-      const data = await boardsService.deleteBoard(req.params.id)
+      const data = await boardsService.deleteBoard({ _id: req.params.id, creatorId: req.userInfo.id })
       return res.send(data)
     } catch (error) {
       next(error)
